Consolidate comparable and graph packages
I accidentally made duplicate Point and TestPoint classes, so I just
"merged" the packages.
